+++ src/esm_runscripts/filelists.py
@@ -25,7 +25,7 @@ def rename_sources_to_targets(config):
     # Purpose of this routine is to make sure that filetype_sources and
     # filetype_targets are set correctly, and _in_work is unset
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             sources = filetype + "_sources" in config[model]
             targets = filetype + "_targets" in config[model]
             in_work = filetype + "_in_work" in config[model]
@@ -127,7 +127,7 @@ def rename_sources_to_targets(config):
 
 def complete_targets(config):
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if filetype + "_sources" in config[model]:
                 for category in config[model][filetype + "_sources"]:
                     if not category in config[model][filetype + "_targets"]:
@@ -161,7 +161,7 @@ def complete_sources(config):
     logger.debug("::: Complete sources")
     helpers.print_datetime(config)
     for filetype in config["general"]["out_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if filetype + "_sources" in config[model]:
                 for category in config[model][filetype + "_sources"]:
                     if not config[model][filetype + "_sources"][category].startswith(
@@ -184,13 +184,13 @@ def reuse_sources(config):
 
     # Put together all the possible reusable file types
     all_reusable_filetypes = []
-    for model in esm_parser.get_components(config, include_system=False):
+    for model in esm_parser.get_components(config, include=["model"]):
         all_reusable_filetypes = list(
             set(all_reusable_filetypes + config[model].get("reusable_filetypes", []))
         )
     # Loop through all the reusable file types
     for filetype in all_reusable_filetypes:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             # Get the model-specific reusable_filetypes, if not existing, get the
             # general ones
             model_reusable_filetypes = config[model].get(
@@ -220,7 +220,7 @@ def choose_needed_files(config):
     # (if exists), and then remove filetype_files
 
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if not filetype + "_files" in config[model]:
                 continue
 
@@ -258,7 +258,7 @@ def choose_needed_files(config):
 
 def globbing(config):
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if filetype + "_sources" in config[model]:
                 # oldconf = copy.deepcopy(config[model])
                 for descr, filename in copy.deepcopy(
@@ -302,7 +302,7 @@ def globbing(config):
 
 def target_subfolders(config):
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if filetype + "_targets" in config[model]:
                 for descr, filename in config[model][filetype + "_targets"].items():
                     # * only in targets if denotes subfolder
@@ -452,7 +452,7 @@ def complete_restart_in(config):
 
 def assemble_intermediate_files_and_finalize_targets(config):
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if filetype + "_targets" in config[model]:
                 if not filetype + "_intermediate" in config[model]:
                     config[model][filetype + "_intermediate"] = {}
@@ -509,7 +509,7 @@ def find_valid_year(config, year):
 
 def replace_year_placeholder(config):
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if filetype + "_targets" in config[model]:
                 if filetype + "_additional_information" in config[model]:
                     for file_category in config[model][
@@ -794,7 +794,7 @@ def log_used_files(config):
     )
 
     with open(flist_file, "w") as flist:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             flist.write(
                 f"These files are used for\n"
                 f"experiment {config['general']['expid']}\n"
@@ -888,7 +888,7 @@ def compute_and_log_file_checksums(config):
     files_not_handled_by_filelists = copy.deepcopy(checksums)
 
     # Loop over all components, file types, and files
-    for component in esm_parser.get_components(config, include_system=False):
+    for component in esm_parser.get_components(config, include=["model"]):
         component_files = {}
         for filetype in filetypes:
             component_config = config[component]
@@ -990,7 +990,7 @@ def check_for_unknown_files(config):
     ]
 
     for filetype in config["general"]["all_model_filetypes"]:
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             if filetype + "_sources" in config[model]:
                 known_files += list(config[model][filetype + "_sources"].values())
                 known_files += list(config[model][filetype + "_targets"].values())
@@ -1183,7 +1183,7 @@ def copy_files(config, filetypes, source, target):
     files_to_be_moved = []
     for filetype in [filetype for filetype in filetypes if not filetype == "ignore"]:
         # Loop through the components
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             # If there is a source of this file type in the model
             if filetype + "_" + text_source in config[model]:
                 this_text_target = text_target
@@ -1426,7 +1426,7 @@ def filter_allowed_missing_files(config):
     for missing_file_source, missing_file_target in missing_files.items():
         missing_file_source_fname = pathlib.Path(missing_file_source).name
         missing_file_target_fname = pathlib.Path(missing_file_target).name
-        for model in esm_parser.get_components(config, include_system=False):
+        for model in esm_parser.get_components(config, include=["model"]):
             for allowed_missing_pattern in config[model].get(
                 "allowed_missing_files", []
             ):
@@ -1510,7 +1510,7 @@ def _check_fesom_missing_files(config):
 
 
 def create_missing_file_movement_entries(config):
-    for model in esm_parser.get_components(config, include_system=False):
+    for model in esm_parser.get_components(config, include=["model"]):
         if not "file_movements" in config[model]:
             config[model]["file_movements"] = {}
         for filetype in config["general"]["all_model_filetypes"] + [
@@ -1587,7 +1587,7 @@ def complete_all_file_movements(config):
 
     config = create_missing_file_movement_entries(config)
 
-    for model in esm_parser.get_components(config, include_system=False):
+    for model in esm_parser.get_components(config, include=["model"]):
         mconfig = config[model]
         if "file_movements" in mconfig:
             for filetype in config["general"]["all_model_filetypes"] + [
@@ -1610,7 +1610,7 @@ def complete_all_file_movements(config):
                             )
                         del mconfig["file_movements"][filetype]["all_directions"]
 
-    for model in esm_parser.get_components(config, include_system=False):
+    for model in esm_parser.get_components(config, include=["model"]):
         mconfig = config[model]
         if "file_movements" in mconfig:
             # Complete movements with general